--- binder/postBuild
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-export JUPYTER_SERVER_URL
-jupyter serverextension enable --py jupyter_server_proxy --sys-prefix
-python setup.py build
-pip install .[all]
