--- binder/requirements.txt
@@ -1,1 +1,0 @@
-jupyter-server-proxy
