+++ requirements/dev.txt
@@ -1,6 +1,4 @@
-jupyterlab
 pre-commit
-notebook
 twine
 wheel
 setuptools_scm
